Document single-decode-thread invariant; fix PSD debug swprintf type mismatch
- MainDlg.cpp: document why NUM_THREADS must stay 1. The format decoders
  (PngReader, JxlReader, AvifReader, WebP, etc.) keep per-codec state in
  class-static caches that are not thread-safe; a single load thread
  serializes all decodes. Raising the count would let two threads decode
  concurrently and race / double-free that shared cache state. Safe today,
  but the invariant was undocumented.
- PSDWrapper.cpp: cast the ptrdiff/__int64 and unsigned args of the
  DEBUG-only swprintf to int to match the %d format specifiers (clears
  C4477 / C6328 size-mismatch).

Also ran MSVC /analyze over the project as an automated bug hunt: no new
memory-safety defects in the decoders - the buffer-overrun flags on
ReaderTGA/QOIWrapper/ResizeFilter are false positives on size-validated,
deterministic code (verified by hand). Remaining /analyze findings are
style nags and false "handle could be NULL" on Win32 calls; left unchanged.

Builds clean (0 errors) in Debug|x64.

+++ src/JPEGView/PSDWrapper.cpp
@@ -378,7 +378,7 @@ CJPEGImage* PsdReader::ReadImage(LPCTSTR strFileName, bool& bOutOfMemory)
 #ifdef DEBUG
 					if (p != pOffset) {
 						WCHAR buf[100];
-						swprintf(buf, _T("Misaligned scan line bytes (%+d) for channel %d row %d\n"), p - pOffset, channel, row);
+						swprintf(buf, _T("Misaligned scan line bytes (%+d) for channel %d row %d\n"), (int)(p - pOffset), (int)channel, (int)row);
 						::OutputDebugString(buf);
 					}
 #endif
